+++ wp_api/keywords/posts/crud.robot
@@ -4,7 +4,6 @@ Library    libraries.authentication.custom_session.CustomSession
 Library    libraries.process_data.process_data.ProcessData
 Library    libraries.assertions.verification.Verification
 Library    libraries.process_data.url_assembler.UrlAssembler
-Library    libraries.assertions.pages_verification.PagesVerification
 Variables     wp_api/resources/data/bodies/posts.py
 
 *** Variables ***
