+++ wp_api/resources/data/bodies/block_body_generator.py
@@ -1,5 +1,6 @@
 class BodyGenerator():
     """generates bodies"""
+    
     def block_body(self, type):
         """return valid bodies"""
         if type == 'Valid' or type == 'valid':
@@ -27,6 +28,7 @@ class BodyGenerator():
             return {'slug': 'edited slug'}
         if type == 'password' or type == 'Password':
             return {'password': 'new password'}
+
     def data_generator(self, type):
         """Returns invalid bodies"""
         if type == 'invalid status' or type == 'Invalid Status':
